backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
python-dotenv==1.0.1
httpx==0.27.0
scikit-learn>=1.6.1
xgboost>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
joblib==1.3.2
reportlab==4.0.7
python-dateutil==2.8.2
google-generativeai==0.8.3
openai==1.65.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
cryptography==41.0.7
